requirements.txt
pandapipes>=0.8.0
pandapower>=2.14.0
geopandas>=0.12.0
networkx>=2.8.0
numpy>=1.21.0
pandas>=1.3.0
shapely>=1.8.0
scipy>=1.7.0
pyproj>=3.3.0
fiona>=1.8.0
matplotlib>=3.5.0
pyyaml>=6.0
questionary>=1.10.0
osmnx>=1.2.0
contextily>=1.4.0
folium>=0.14.0
openai>=1.0.0
adk>=0.1.0
